New feature for Members. Closed#723
Added new Member Feature in Admin Panel in Settings (Keep uploaded content after Memeber deleted).
If enabled it, than after deleting a Member all uploaded images of deleted user will saved on upload folder.
If disable, all uploaded stuff will be deleted.

+++ includes/classes/ia.core.users.php
@@ -333,9 +333,11 @@ SQL;
                 // delete associated auth providers
                 $this->iaDb->delete(iaDb::convertIds($entry['id'], 'member_id'), self::$_providersTable);
 
-                // delete member uploads folder
-                $folder = IA_UPLOADS . iaUtil::getAccountDir($entry['username']);
-                iaUtil::cascadeDeleteFiles($folder, true) && @rmdir($folder);
+                if (!$this->iaCore->get('removed_members_keep_images')){
+                    // delete member uploads folder
+                    $folder = IA_UPLOADS . iaUtil::getAccountDir($entry['username']);
+                    iaUtil::cascadeDeleteFiles($folder, true) && @rmdir($folder);
+                }
 
                 $iaLog->write(iaLog::ACTION_DELETE,
                     ['item' => 'member', 'name' => $entry['fullname'], 'id' => $entry['id']]);
